+++ src/main.rs
@@ -66,6 +66,7 @@ async fn main() -> std::io::Result<()> {
             .configure(api::saves::config)
             .configure(api::game_settings::config)
             .configure(api::events::config)
+            .configure(api::feedback::config)
             .configure(api::socket_ticket::config)
             .route("/ws", web::get().to(sockets::ws::ws_index));
         if let Some(p) = pool.clone() {